requirements.txt
apscheduler
Flask
Flask-SQLAlchemy
mysql-replication
pymysql
python-dotenv
requests